Convert `MatchmakingSelectPanel` card content classes into nested classes
